requirements.txt
apache-bookkeeper-client==4.12.0
black==19.10b0
coloredlogs==14.0
invoke==1.4.1
protobuf==3.14.0
pulsar-client==2.6.1
